Move Explorer, Types and Utils to base package
