+++ .gitignore
@@ -5,3 +5,4 @@ packages/**/coverage
 packages/**/dist
 package-lock.json
 tsconfig.tsbuildinfo
+tsconfig.build.tsbuildinfo
