+++ package.json
@@ -5,7 +5,7 @@
     "packages/*"
   ],
   "scripts": {
-    "clean": "git clean -Xdf -e '!node_modules' .",
+    "clean": "npm run clean --workspaces",
     "build": "npm run build --workspaces",
     "test": "npm run test --workspaces",
     "lint": "npm run lint --workspaces",
