+++ packages/base/package.json
@@ -1,7 +1,6 @@
 {
-  "name": "cosmiconfig",
+  "name": "@cosmiconfig/base",
   "version": "9.0.0",
-  "description": "Find and load configuration from a package.json property, rc file, TypeScript module, and more!",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
@@ -15,7 +14,6 @@
     "dev": "npm run build:tsc -- --watch",
     "lint": "eslint --ext .js,.ts .",
     "lint:fix": "eslint --ext .js,.ts . --fix",
-    "lint:md": "remark-preset-davidtheclark",
     "test": "vitest run --coverage",
     "test:watch": "vitest"
   },
@@ -41,23 +39,7 @@
     "url": "https://github.com/cosmiconfig/cosmiconfig/issues"
   },
   "homepage": "https://github.com/cosmiconfig/cosmiconfig#readme",
-  "peerDependencies": {
-    "typescript": ">=4.9.5"
-  },
-  "peerDependenciesMeta": {
-    "typescript": {
-      "optional": true
-    }
-  },
   "dependencies": {
-    "@cosmiconfig/base": "9.0.0",
-    "import-fresh": "^3.3.0",
-    "js-yaml": "^4.1.0",
-    "parse-json": "^5.2.0"
-  },
-  "devDependencies": {
-    "@types/js-yaml": "^4.0.5",
-    "@types/parse-json": "^4.0.0",
     "env-paths": "^2.2.1"
   },
   "engines": {
