+++ packages/base/src/defaults.ts
@@ -1,0 +1,21 @@
+// TODO make them configurable via options
+
+export const globalConfigSearchPlaces = [
+  'config',
+  'config.json',
+  'config.yaml',
+  'config.yml',
+  'config.js',
+  'config.ts',
+  'config.cjs',
+  'config.mjs',
+];
+export const globalConfigSearchPlacesSync = [
+  'config',
+  'config.json',
+  'config.yaml',
+  'config.yml',
+  'config.js',
+  'config.ts',
+  'config.cjs',
+];
