+++ packages/base/test/getPropertyByPath.test.ts
