+++ packages/cosmiconfig/src/loaders.ts
@@ -4,7 +4,7 @@ import { existsSync, rmSync, writeFileSync } from 'fs';
 import { rm, writeFile } from 'fs/promises';
 import path from 'path';
 import { pathToFileURL } from 'url';
-import { Loader, LoaderSync } from './types.js';
+import { Loader, LoaderSync } from '@cosmiconfig/base';
 
 let importFresh: typeof import('import-fresh');
 export const loadJsSync: LoaderSync = function loadJsSync(filepath) {
